refactor main.py to reorganise api-related code vs business logic and background worker tasks
