+++ src/api/main.py
@@ -1,315 +1,45 @@
-"""Definition of endpoints/routers for the webserver."""
-
-# pylint: disable=too-many-lines
+"""FastAPI application entry point."""
 
 from __future__ import annotations
 
-import asyncio
-import json
 import logging
-import platform
-import shutil
-import tarfile
-from collections.abc import AsyncGenerator, Iterable
+from collections.abc import AsyncGenerator
 from contextlib import asynccontextmanager
-from datetime import datetime, timedelta
-from pathlib import Path
-from typing import Annotated, Any, cast
 
-import requests
-from fastapi import BackgroundTasks, Depends, FastAPI, HTTPException, Security, status
-from sqlmodel import Session, SQLModel, create_engine, select
+from fastapi import FastAPI
 
-from api.activescale import (
-    download_file_to_disk,
-    get_activescale_client_context,
-    init_activescale,
-    initiate_object_restore,
-    is_object_ready_for_download,
-    object_exists,
-    upload_file,
-)
-from api.archive_reassembly import (
-    load_archive_manifest,
-    ordered_part_object_keys,
-    reassemble_archive_from_manifest,
-)
-from api.archive_chunks import build_chunked_tar_archive
 from api.cors import add_cors_middleware
-from api.manifests import bag_directory, bagit_exists, create_manifests_directory, validate_bag
-from api.security import ApiKey, validate_api_key, validate_permissions
-from config import get_settings
-from crate.ro_builder import ROBuilder
-from crate.ro_loader import ROLoader
-from models.common import ResearchDriveName, calculate_retention_end_date
-from models.request import CreateRetrievalRequest, CreateSubmissionRequest
-from models.response import (
-    CodeResponse,
-    CreateRetrievalResponse,
-    CreateSubmissionResponse,
-    DriveInfoResponse,
-    DriveResponse,
-    ErrorResponse,
-    MemberResponse,
-    PersonResponse,
-    ProjectResponse,
-    RoleResponse,
-    SubmissionResponse,
-)
-from models.retrieval import (
-    ACTIVE_RETRIEVAL_STAGES,
-    ArchiveRetrieval,
-    RetrievalJobStage,
-)
-from models.submission import (
-    ACTIVE_STAGES,
-    RETRYABLE_STAGES,
-    ArchiveSubmission,
-    JobStage,
-)
-from service.projectdb import get_projectdb_client, init_projectdb
-from service.projectdb_client import ProjectDBClient
+from api.dependencies import create_db_and_tables, engine
+from api.routers import drives, retrievals, submissions
+from service.activescale import init_activescale
+from service.projectdb import init_projectdb
+from utils.logging import configure_logging, log_event
+from utils.paths import validate_archive_path_configuration
+from workers.retrieval_worker import _reconcile_interrupted_retrievals
+from workers.submission_worker import _reconcile_interrupted_jobs
 
+configure_logging()
 
-def _configure_logging() -> None:
-    """Initialize logging once and align root level with configured settings."""
-    root_logger = logging.getLogger()
-    configured_level = get_settings().log_level
-    if not root_logger.handlers:
-        logging.basicConfig(
-            level=configured_level,
-            format="%(asctime)s [%(name)s] %(levelname)s: %(message)s",
-        )
-    else:
-        root_logger.setLevel(configured_level)
-
-
-_configure_logging()
-logger = logging.getLogger(__name__)
-
-
-def _log_event(
-    level: int,
-    event: str,
-    *,
-    exc_info: bool = False,
-    **context: Any,
-) -> None:
-    payload = {"event": event, **context}
-    logger.log(level, json.dumps(payload, default=str), exc_info=exc_info)
-
-
-def _elapsed_ms(started_at: datetime) -> int:
-    """Compute elapsed milliseconds from a start timestamp."""
-    return int((datetime.now() - started_at).total_seconds() * 1000)
-
-
-def _is_windows_runtime() -> bool:
-    """Return True when running on Windows.
-
-    Uses platform.system() instead of os.name to avoid some static analyzers
-    folding branches as unreachable based on host/editor platform settings.
-    """
-    return platform.system().lower().startswith("win")
-
-
-def _validate_archive_path_configuration() -> None:
-    """Validate archive path configuration at startup.
-
-    On non-Windows with UNC SMB base paths, a local mount base is required so
-    bagit/rocrate can perform filesystem operations.
-    """
-    settings = get_settings()
-    smb_base = settings.smb_drive_base_path.strip()
-    if _is_windows_runtime() or not smb_base.startswith("//"):
-        return
-
-    mount_base = settings.smb_linux_mount_base_path.strip()
-    if not mount_base:
-        raise RuntimeError(
-            "SMB_LINUX_MOUNT_BASE_PATH is required on Linux when "
-            "SMB_DRIVE_BASE_PATH is a UNC path."
-        )
-
-    mount_path = Path(mount_base)
-    if not mount_path.exists() or not mount_path.is_dir():
-        raise RuntimeError(
-            "Configured SMB_LINUX_MOUNT_BASE_PATH is invalid: "
-            f"{mount_path}. Ensure the CIFS mount parent exists."
-        )
-
-    _log_event(
-        logging.INFO,
-        "startup.archive_path_config_validated",
-        smb_drive_base_path=smb_base,
-        smb_linux_mount_base_path=str(mount_path),
-    )
-
-
-def _resolve_drive_path_for_archive(drive_name: str) -> Path:
-    """Resolve a filesystem path usable by bagit/rocrate for archive operations.
-
-    On Linux, UNC paths (//server/share) are not valid local filesystem targets.
-    In that case, require SMB_LINUX_MOUNT_BASE_PATH and resolve to:
-    <mount_base>/<drive_name>.
-    """
-    settings = get_settings()
-    smb_base = settings.smb_drive_base_path.strip()
-    drive_path = Path(smb_base) / drive_name
-
-    if _is_windows_runtime() or not str(drive_path).startswith("//"):
-        return drive_path
-
-    mount_base = settings.smb_linux_mount_base_path.strip()
-    if not mount_base:
-        raise RuntimeError(
-            "SMB_LINUX_MOUNT_BASE_PATH is required on Linux when "
-            "SMB_DRIVE_BASE_PATH is a UNC path."
-        )
-
-    mounted_drive_path = Path(mount_base) / drive_name
-    if not mounted_drive_path.exists():
-        raise FileNotFoundError(
-            "Configured mounted drive path does not exist: "
-            f"{mounted_drive_path}. Ensure the CIFS mount is available."
-        )
-    return mounted_drive_path
-
-
-def _validate_archive_path_access(drive_name: str) -> Path:
-    """Validate drive path can be read and written before scheduling archive job.
-
-    Synchronous validation to be called from submission endpoints (before 201).
-    Resolves the archive path and validates read/write permissions early.
-    """
-    drive_path = _resolve_drive_path_for_archive(drive_name)
-
-    if not drive_path.exists() or not drive_path.is_dir():
-        raise FileNotFoundError(
-            f"Drive path does not exist or is not a directory: {drive_path}"
-        )
-
-    # Read probe to validate source permissions
-    try:
-        _ = next(drive_path.iterdir(), None)
-    except Exception as e:
-        raise PermissionError(f"Cannot read source directory {drive_path}: {e}") from e
-
-    # Write probe to validate output permissions
-    probe_file = drive_path / ".driveoff_write_probe"
-    try:
-        with open(probe_file, "wb") as f:
-            f.write(b"ok")
-        probe_file.unlink(missing_ok=True)
-    except Exception as e:
-        raise PermissionError(f"Cannot write to directory {drive_path}: {e}") from e
-
-    # Validate local temp base is writable for generated archive artifacts.
-    temp_base = Path(get_settings().archive_temp_base_path).expanduser()
-    try:
-        temp_base.mkdir(parents=True, exist_ok=True)
-        temp_probe = temp_base / ".driveoff_temp_probe"
-        with open(temp_probe, "wb") as f:
-            f.write(b"ok")
-        temp_probe.unlink(missing_ok=True)
-    except Exception as e:
-        raise PermissionError(
-            f"Cannot write to local archive temp base {temp_base}: {e}"
-        ) from e
-
-    return drive_path
-
-
-def _resolve_archive_output_location(drive_name: str) -> Path:
-    """Resolve local output directory for generated archive artifacts."""
-    temp_base = Path(get_settings().archive_temp_base_path).expanduser()
-    safe_drive_name = drive_name.replace("/", "_").replace("\\", "_")
-    return temp_base / "bagit_temp" / safe_drive_name
-
-
-def _validate_destination_path(destination_path: str) -> Path:
-    """Validate that a destination path for archive retrieval exists and is writable.
-
-    Raises FileNotFoundError if the path does not exist or is not a directory,
-    and PermissionError if write access is denied.
-    """
-    dest = Path(destination_path)
-    if not dest.exists() or not dest.is_dir():
-        raise FileNotFoundError(
-            f"Destination path does not exist or is not a directory: {dest}"
-        )
-    probe_file = dest / ".driveoff_dest_probe"
-    try:
-        with open(probe_file, "wb") as f:
-            f.write(b"ok")
-        probe_file.unlink(missing_ok=True)
-    except Exception as e:
-        raise PermissionError(f"Cannot write to destination path {dest}: {e}") from e
-    return dest
-
-
-def _as_bad_request_for_archive_path(
-    drive_name: str,
-    error: FileNotFoundError | PermissionError | RuntimeError,
-) -> HTTPException:
-    """Convert archive path validation errors into consistent client errors."""
-    _log_event(
-        logging.WARNING,
-        "submission.archive_path_validation_failed",
-        drive_name=drive_name,
-        error=str(error),
-        error_type=type(error).__name__,
-    )
-    return HTTPException(
-        status_code=status.HTTP_400_BAD_REQUEST,
-        detail=(f"Archive path validation failed for drive {drive_name}: {error}"),
-    )
-
-
-# Ensure driveoff directory is created
-(Path.home() / ".driveoff").mkdir(exist_ok=True)
-DB_FILE_NAME = Path.home() / ".driveoff" / "database.db"
-DB_URL = f"sqlite:///{DB_FILE_NAME}"
-
-connect_args = {"check_same_thread": False}
-engine = create_engine(DB_URL, connect_args=connect_args, echo=False)
-
-
-def create_db_and_tables() -> None:
-    """Create database structure for archive submissions and manifests."""
-    SQLModel.metadata.create_all(engine)
-
-
-def get_session() -> Iterable[Session]:
-    """Return a Session object."""
-    with Session(engine) as session:
-        yield session
-
-
-SessionDep = Annotated[Session, Depends(get_session)]
-ProjectDbDep = Annotated[ProjectDBClient, Depends(get_projectdb_client)]
+ENDPOINT_PREFIX = "/api/v1"
 
 
 @asynccontextmanager
 async def lifespan(app_instance: FastAPI) -> AsyncGenerator[None, None]:
-    """Lifecycle method for the API
+    """Lifecycle method for the API.
 
-    This creates DB tables and initialises the ActiveScale, ProjectDB,
-    and research drive service clients during application startup so
-    routes can depend on them.
-
-    It also performs reconciliation of any active archive jobs that were in-flight
-    when the process last exited, marking them as 'abandoned' so they can be retried.
+    Creates DB tables and initialises service clients during application startup
+    so routes can depend on them.  Also reconciles any archive jobs that were
+    in-flight when the process last exited so operators can retry them.
     """
     create_db_and_tables()
-    _validate_archive_path_configuration()
+    validate_archive_path_configuration()
     init_projectdb(app_instance)
     init_activescale(app_instance)
 
     try:
         _reconcile_interrupted_jobs()
     except Exception as e:  # pylint: disable=broad-exception-caught
-        _log_event(
+        log_event(
             logging.WARNING,
             "startup.reconciliation_failed",
             error=str(e),
@@ -319,7 +49,7 @@ async def lifespan(app_instance: FastAPI) -> AsyncGenerator[None, None]:
     try:
         _reconcile_interrupted_retrievals()
     except Exception as e:  # pylint: disable=broad-exception-caught
-        _log_event(
+        log_event(
             logging.WARNING,
             "startup.retrieval_reconciliation_failed",
             error=str(e),
@@ -330,1865 +60,10 @@ async def lifespan(app_instance: FastAPI) -> AsyncGenerator[None, None]:
     engine.dispose()
 
 
-def _reconcile_interrupted_jobs() -> None:
-    """Mark any jobs that were active when the process last exited as abandoned.
-
-    FastAPI BackgroundTasks are volatile; if the process restarts while a job
-    is in an active stage (queued/running/uploading/cleanup) that work is lost.
-    We surface this to operators as 'abandoned' so they can trigger a manual retry.
-    """
-    with Session(engine) as session:
-        active_stage_values = [s.value for s in ACTIVE_STAGES]
-        stage_column = cast(Any, ArchiveSubmission.stage)
-        interrupted = session.exec(
-            select(ArchiveSubmission).where(stage_column.in_(active_stage_values))
-        ).all()
-        if not interrupted:
-            return
-        now = datetime.now()
-        for submission in interrupted:
-            previous_stage = submission.stage
-            _log_event(
-                logging.WARNING,
-                "submission.abandoned_on_startup",
-                submission_id=submission.id,
-                drive_name=submission.drive_name,
-                previous_stage=previous_stage,
-            )
-            submission.stage = JobStage.ABANDONED
-            submission.failure_reason = (
-                f"Process restarted while job was in stage '{previous_stage.value}'."
-                " Retry this job to resume."
-            )
-            submission.failed_timestamp = now
-            submission.last_updated_timestamp = now
-            session.add(submission)
-        session.commit()
-        _log_event(
-            logging.WARNING,
-            "startup.reconciliation_complete",
-            abandoned_count=len(interrupted),
-        )
-
-
-def _reconcile_interrupted_retrievals() -> None:
-    """Mark any retrieval jobs that were active when the process last exited as failed.
-
-    FastAPI BackgroundTasks are volatile; if the process restarts while a retrieval
-    is in an active stage the work is permanently lost.  We surface this as 'failed'
-    so operators know to re-submit a retrieval request.
-    """
-    with Session(engine) as session:
-        active_stage_values = [s.value for s in ACTIVE_RETRIEVAL_STAGES]
-        stage_column = cast(Any, ArchiveRetrieval.stage)
-        interrupted = session.exec(
-            select(ArchiveRetrieval).where(stage_column.in_(active_stage_values))
-        ).all()
-        if not interrupted:
-            return
-        now = datetime.now()
-        for retrieval in interrupted:
-            previous_stage = retrieval.stage
-            _log_event(
-                logging.WARNING,
-                "retrieval.abandoned_on_startup",
-                retrieval_id=retrieval.id,
-                drive_name=retrieval.drive_name,
-                previous_stage=previous_stage.value,
-            )
-            retrieval.stage = RetrievalJobStage.FAILED
-            retrieval.failure_reason = (
-                f"Process restarted while retrieval was in stage '{previous_stage.value}'."
-                " Re-submit a retrieval request to restart."
-            )
-            retrieval.failed_timestamp = now
-            retrieval.last_updated_timestamp = now
-            session.add(retrieval)
-        session.commit()
-        _log_event(
-            logging.WARNING,
-            "startup.retrieval_reconciliation_complete",
-            failed_count=len(interrupted),
-        )
-
-
 app = FastAPI(lifespan=lifespan, title="Research Drive Archive API", version="1.0.0")
 
-# Send CORS headers to enable frontend to contact API.
 add_cors_middleware(app)
 
-ENDPOINT_PREFIX = "/api/v1"
-
-
-@app.get(
-    ENDPOINT_PREFIX + "/driveinfo",
-    response_model=DriveInfoResponse,
-    responses={
-        401: {"model": ErrorResponse, "description": "Invalid or missing API key"},
-        404: {"model": ErrorResponse, "description": "Drive or project not found"},
-        500: {"model": ErrorResponse, "description": "Internal server error"},
-    },
-)
-async def get_drive_info(
-    drive_name: ResearchDriveName,
-    projectdb: ProjectDbDep,
-    api_key: ApiKey = Security(validate_api_key),
-) -> DriveInfoResponse:
-    """Retrieve drive and project info from ProjectDB for display.
-
-    Looks up the drive by name, resolves the associated project,
-    and returns combined info including members and codes.
-    """
-    validate_permissions("GET", api_key)
-
-    # Look up drive in ProjectDB
-    try:
-        drive_data = projectdb.get_research_drive_by_name(drive_name)
-        if not drive_data:
-            raise HTTPException(
-                status_code=404,
-                detail=f"Research Drive {drive_name} not found in ProjectDB.",
-            )
-        if isinstance(drive_data, list):
-            drive_data = drive_data[0]
-    except (requests.RequestException, ValueError) as e:
-        raise HTTPException(
-            status_code=status.HTTP_502_BAD_GATEWAY,
-            detail=(
-                "ProjectDB request failed while fetching drive"
-                f" {drive_name}: {str(e)}"
-            ),
-        ) from e
-
-    # Resolve project from drive
-    try:
-        drive_projects = projectdb.get_research_drive_projects(
-            drive_data["id"], expand=["project"]
-        )
-    except (requests.RequestException, ValueError) as e:
-        raise HTTPException(
-            status_code=status.HTTP_502_BAD_GATEWAY,
-            detail=(
-                "ProjectDB request failed while fetching projects"
-                f" for drive {drive_name}: {str(e)}"
-            ),
-        ) from e
-
-    if not drive_projects or len(drive_projects) == 0:
-        raise HTTPException(
-            status_code=404,
-            detail=f"No projects associated with drive {drive_name}",
-        )
-
-    # Use the first project (single-project drives are the common case)
-    project_id = drive_projects[0]["project"]["id"]
-
-    # Fetch full project data with codes and services
-    try:
-        project_data = projectdb.get_project(
-            pid=project_id,
-            expand=["codes", "status", "services"],
-        )
-    except (requests.RequestException, ValueError) as e:
-        raise HTTPException(
-            status_code=status.HTTP_502_BAD_GATEWAY,
-            detail=(
-                "ProjectDB request failed while fetching project"
-                f" {project_id}: {str(e)}"
-            ),
-        ) from e
-
-    # Fetch members
-    try:
-        members_raw = projectdb.get_project_members(
-            project_id,
-            expand=["person", "role", "person.identities", "person.status"],
-        )
-        members_raw = filter_member_identities(members_raw)
-    except (requests.RequestException, ValueError) as e:
-        raise HTTPException(
-            status_code=status.HTTP_502_BAD_GATEWAY,
-            detail=(
-                "ProjectDB request failed while fetching members"
-                f" for project {project_id}: {str(e)}"
-            ),
-        ) from e
-
-    # Build drive response, preferring service-level data (has first_day/last_day)
-    drive_service = None
-    for rd in project_data.get("services", {}).get("research_drive", []):
-        if rd.get("name") == drive_name:
-            drive_service = rd
-            break
-
-    drive_resp = DriveResponse(
-        id=drive_data["id"],
-        name=drive_name,
-        allocated_gb=drive_data["allocated_gb"],
-        used_gb=drive_data["used_gb"],
-        free_gb=drive_data["free_gb"],
-        percentage_used=drive_data["percentage_used"],
-        date=drive_data["date"],
-        first_day=drive_service.get("first_day") if drive_service else None,
-        last_day=drive_service.get("last_day") if drive_service else None,
-    )
-
-    # Build project response
-    codes = _build_codes(project_data)
-    members = _build_members(members_raw)
-    project_resp = ProjectResponse(
-        id=project_data["id"],
-        title=project_data.get("title", ""),
-        description=project_data.get("description", ""),
-        division=project_data.get("division", ""),
-        start_date=project_data.get("start_date", ""),
-        end_date=project_data.get("end_date", ""),
-        codes=codes,
-        members=members,
-    )
-
-    return DriveInfoResponse(drive=drive_resp, project=project_resp)
-
-
-@app.post(
-    ENDPOINT_PREFIX + "/submission",
-    status_code=status.HTTP_201_CREATED,
-    response_model=CreateSubmissionResponse,
-    responses={
-        400: {"model": ErrorResponse, "description": "Invalid submission request"},
-        401: {"model": ErrorResponse, "description": "Invalid or missing API key"},
-        404: {"model": ErrorResponse, "description": "Drive or project not found"},
-        409: {
-            "model": ErrorResponse,
-            "description": "Drive has already been archived",
-        },
-        422: {"description": "Validation error"},
-        500: {"model": ErrorResponse, "description": "Internal server error"},
-        502: {
-            "model": ErrorResponse,
-            "description": "ProjectDB upstream request failed",
-        },
-    },
-)
-async def create_submission(
-    request: CreateSubmissionRequest,
-    session: SessionDep,
-    background_tasks: BackgroundTasks,
-    projectdb: ProjectDbDep,
-    api_key: ApiKey = Security(validate_api_key),
-) -> CreateSubmissionResponse:
-    """Create a new archive submission for a research drive.
-
-    Validates drive exists in ProjectDB, resolves project_id if needed,
-    and schedules RO-Crate generation as a background task.
-    """
-    validate_permissions("POST", api_key)
-    existing_submission = session.exec(
-        select(ArchiveSubmission).where(
-            ArchiveSubmission.drive_name == request.drive_name
-        )
-    ).first()
-
-    if (
-        existing_submission
-        and existing_submission.stage == JobStage.COMPLETED
-        and not request.force
-    ):
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail=(
-                f"Drive {request.drive_name} has already been successfully archived."
-                " Use force=true to re-run."
-            ),
-        )
-
-    if (
-        existing_submission
-        and existing_submission.stage == JobStage.COMPLETED
-        and request.force
-    ):
-        _log_event(
-            logging.WARNING,
-            "submission.force_rerun",
-            drive_name=request.drive_name,
-            previous_stage=existing_submission.stage.value,
-            submission_id=existing_submission.id,
-        )
-
-    if existing_submission and existing_submission.stage in ACTIVE_STAGES:
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail=(
-                f"Drive {request.drive_name} already has an active archive job"
-                f" in stage '{existing_submission.stage.value}'."
-                " Wait for it to finish or use the retry endpoint if it has failed."
-            ),
-        )
-
-    try:
-        drive = _validate_drive(projectdb, request.drive_name)
-        project_id = _resolve_project_id(projectdb, drive, request)
-
-        # Validate drive paths are accessible before creating submission record
-        # This ensures we fail fast with clear errors before returning 201
-        _validate_archive_path_access(request.drive_name)
-
-        submission = _upsert_submission(
-            session,
-            existing_submission,
-            drive,
-            project_id,
-            request,
-        )
-
-        if submission.id is None:
-            raise HTTPException(
-                status_code=500,
-                detail="Archive submission record is missing ID after creation.",
-            )
-
-        # Schedule async RO-Crate generation
-        _log_event(
-            logging.INFO,
-            "submission.background_task_scheduled",
-            drive_name=request.drive_name,
-            submission_id=submission.id,
-        )
-
-        background_tasks.add_task(
-            generate_ro_crate,
-            drive,
-            submission.id,
-            projectdb_client=projectdb,
-        )
-
-        status_word = "updated" if existing_submission else "created"
-        return CreateSubmissionResponse(
-            message=(
-                f"Archive submission {status_word}"
-                f" for {request.drive_name}."
-                " RO-Crate generation and upload to ActiveScale is in progress."
-            )
-        )
-    except (FileNotFoundError, PermissionError, RuntimeError) as e:
-        raise _as_bad_request_for_archive_path(request.drive_name, e) from e
-    except HTTPException:
-        raise
-    except Exception as e:
-        _log_event(
-            logging.ERROR,
-            "submission.create.unexpected_error",
-            drive_name=request.drive_name,
-            error=str(e),
-            error_type=type(e).__name__,
-            exc_info=True,
-        )
-        raise HTTPException(
-            status_code=500,
-            detail=(
-                "An error occurred while creating archive submission"
-                f" for {request.drive_name}."
-            ),
-        ) from e
-
-
-@app.post(
-    ENDPOINT_PREFIX + "/submission/{drive_name}/retry",
-    status_code=status.HTTP_200_OK,
-    response_model=CreateSubmissionResponse,
-    responses={
-        401: {"model": ErrorResponse, "description": "Invalid or missing API key"},
-        404: {"model": ErrorResponse, "description": "No submission found for drive"},
-        409: {
-            "model": ErrorResponse,
-            "description": "Job is active or already completed",
-        },
-        500: {"model": ErrorResponse, "description": "Internal server error"},
-    },
-)
-async def retry_submission(
-    drive_name: ResearchDriveName,
-    session: SessionDep,
-    background_tasks: BackgroundTasks,
-    projectdb: ProjectDbDep,
-    api_key: ApiKey = Security(validate_api_key),
-    force: bool = False,
-) -> (
-    CreateSubmissionResponse
-):  # pylint: disable=too-many-arguments,too-many-positional-arguments
-    """Retry a failed or abandoned archive job for a research drive."""
-    validate_permissions("POST", api_key)
-
-    submission = session.exec(
-        select(ArchiveSubmission).where(ArchiveSubmission.drive_name == drive_name)
-    ).first()
-
-    if submission is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"No archive submission found for drive {drive_name}.",
-        )
-
-    if submission.stage in ACTIVE_STAGES:
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail=(
-                f"Drive {drive_name} already has an active archive job"
-                f" in stage '{submission.stage.value}'."
-                " Wait for it to finish."
-            ),
-        )
-
-    if submission.stage == JobStage.COMPLETED and not force:
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail=(
-                f"Drive {drive_name} has already been successfully archived."
-                " Use force=true to re-run."
-            ),
-        )
-
-    if submission.stage not in RETRYABLE_STAGES and not (
-        force and submission.stage == JobStage.COMPLETED
-    ):
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail=(
-                f"Drive {drive_name} submission is in stage '{submission.stage.value}'"
-                " which cannot be retried."
-            ),
-        )
-
-    if force and submission.stage == JobStage.COMPLETED:
-        _log_event(
-            logging.WARNING,
-            "submission.force_rerun",
-            drive_name=drive_name,
-            previous_stage=submission.stage.value,
-            submission_id=submission.id,
-        )
-
-    drive = _validate_drive(projectdb, drive_name)
-
-    # Validate drive paths are accessible before mutating submission state.
-    try:
-        _validate_archive_path_access(drive_name)
-    except (FileNotFoundError, PermissionError, RuntimeError) as e:
-        raise _as_bad_request_for_archive_path(drive_name, e) from e
-
-    # Reset to QUEUED for the new attempt
-    now = datetime.now()
-    submission.stage = JobStage.QUEUED
-    submission.failure_reason = None
-    submission.failed_timestamp = None
-    submission.retry_count = (submission.retry_count or 0) + 1
-    submission.last_updated_timestamp = now
-    submission.completed_timestamp = None
-    submission.cleanup_succeeded = None
-    submission.cleanup_error = None
-    session.add(submission)
-    session.commit()
-    session.refresh(submission)
-
-    _log_event(
-        logging.INFO,
-        "submission.retry_scheduled",
-        drive_name=drive_name,
-        submission_id=submission.id,
-        retry_count=submission.retry_count,
-    )
-
-    if submission.id is None:
-        raise HTTPException(
-            status_code=500,
-            detail="Archive submission record is missing ID.",
-        )
-
-    background_tasks.add_task(
-        generate_ro_crate,
-        drive,
-        submission.id,
-        projectdb_client=projectdb,
-    )
-
-    return CreateSubmissionResponse(
-        message=(
-            f"Archive job for {drive_name} queued for retry"
-            f" (attempt {submission.retry_count})."
-            " RO-Crate generation and upload to ActiveScale is in progress."
-        )
-    )
-
-
-def _validate_drive(projectdb: ProjectDBClient, drive_name: str) -> Any:
-    """Fetch and validate drive from ProjectDB."""
-    drive = projectdb.get_research_drive_by_name(drive_name)
-    if not drive:
-        raise HTTPException(
-            status_code=404,
-            detail=f"Research Drive {drive_name} not found in ProjectDB.",
-        )
-    if isinstance(drive, list):
-        drive = drive[0]
-    if not drive:
-        raise HTTPException(
-            status_code=404,
-            detail=f"Research Drive {drive_name} not found in ProjectDB.",
-        )
-    return drive
-
-
-def _resolve_project_id(
-    projectdb: ProjectDBClient,
-    drive: dict[str, Any],
-    request: CreateSubmissionRequest,
-) -> Any:
-    """Resolve the project_id for a drive from ProjectDB."""
-    try:
-        drive_projects = projectdb.get_research_drive_projects(
-            drive["id"], expand=["project"]
-        )
-    except (requests.RequestException, ValueError) as e:
-        raise HTTPException(
-            status_code=status.HTTP_502_BAD_GATEWAY,
-            detail=(
-                "ProjectDB request failed while fetching projects"
-                f" for drive {request.drive_name}: {e}"
-            ),
-        ) from e
-
-    if not drive_projects:
-        raise HTTPException(
-            status_code=404,
-            detail=f"No projects associated with drive {request.drive_name}",
-        )
-
-    if len(drive_projects) == 1:
-        return drive_projects[0]["project"]["id"]
-
-    # Multiple projects – need project_id from request to disambiguate
-    if request.project_id:
-        for dp in drive_projects:
-            if dp["project"]["id"] == request.project_id:
-                return request.project_id
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=(
-                f"Multiple projects associated with drive {request.drive_name}. "
-                "Please provide project_id to disambiguate."
-            ),
-        )
-
-    raise HTTPException(
-        status_code=400,
-        detail="Could not determine project_id for archive",
-    )
-
-
-def _upsert_submission(
-    session: Session,
-    existing_submission: ArchiveSubmission | None,
-    drive: dict[str, Any],
-    project_id: int,
-    request: CreateSubmissionRequest,
-) -> ArchiveSubmission:
-    """Update an existing pending submission or create a new one."""
-    if existing_submission:
-        existing_submission.drive_id = drive["id"]
-        existing_submission.project_id = project_id
-        existing_submission.retention_period_years = request.retention_period_years
-        existing_submission.retention_period_justification = (
-            request.retention_period_justification
-        )
-        existing_submission.data_classification = request.data_classification
-        existing_submission.failure_reason = None
-        existing_submission.failed_timestamp = None
-        existing_submission.archive_file_key = None
-        existing_submission.archive_object_prefix = None
-        existing_submission.archive_manifest_key = None
-        existing_submission.archive_part_keys_json = None
-        existing_submission.archive_part_count = None
-        existing_submission.archive_total_bytes = None
-        submission = existing_submission
-    else:
-        submission = ArchiveSubmission(
-            drive_id=drive["id"],
-            project_id=project_id,
-            drive_name=request.drive_name,
-            retention_period_years=request.retention_period_years,
-            retention_period_justification=request.retention_period_justification,
-            data_classification=request.data_classification,
-        )
-
-    now = datetime.now()
-    submission.stage = JobStage.QUEUED
-    submission.started_timestamp = now
-    submission.last_updated_timestamp = now
-    submission.completed_timestamp = None
-    submission.cleanup_succeeded = None
-    submission.cleanup_error = None
-
-    session.add(submission)
-    session.commit()
-    session.refresh(submission)
-    if submission.id is None:
-        raise HTTPException(
-            status_code=500,
-            detail="Failed to create archive submission record.",
-        )
-    return submission
-
-
-def _cleanup_job_artifacts(
-    drive_name: str, output_location: Path | None
-) -> tuple[bool, str | None]:
-    """Delete generated local archive artifacts for a submission.
-
-    This intentionally only removes generated artifacts in the archive output
-    area and does not remove source drive content.
-    """
-    if output_location is None:
-        _log_event(
-            logging.INFO,
-            "submission.cleanup.skipped",
-            drive_name=drive_name,
-            reason="no output path resolved",
-        )
-        return True, None
-
-    if not output_location.exists():
-        _log_event(
-            logging.INFO,
-            "submission.cleanup.skipped",
-            drive_name=drive_name,
-            output_location=str(output_location),
-            reason="path does not exist",
-        )
-        return True, None
-
-    try:
-        if output_location.is_file():
-            output_location.unlink()
-        else:
-            shutil.rmtree(output_location)
-    except OSError as e:  # pragma: no cover - best effort cleanup
-        cleanup_error = f"{output_location}: {e}"
-        _log_event(
-            logging.WARNING,
-            "submission.cleanup.failed",
-            drive_name=drive_name,
-            output_location=str(output_location),
-            cleanup_error=cleanup_error,
-        )
-        return False, cleanup_error
-
-    _log_event(
-        logging.INFO,
-        "submission.cleanup.completed",
-        drive_name=drive_name,
-        output_location=str(output_location),
-    )
-    return True, None
-
-
-def _parse_uploaded_part_keys(part_keys_json: str | None) -> list[str]:
-    """Decode persisted uploaded part keys with defensive validation."""
-    if not part_keys_json:
-        return []
-    try:
-        parsed = json.loads(part_keys_json)
-    except json.JSONDecodeError:
-        return []
-    if not isinstance(parsed, list):
-        return []
-    return [str(item) for item in parsed if isinstance(item, str)]
-
-
-def _persist_uploaded_part_keys(
-    session: Session,
-    submission: ArchiveSubmission,
-    uploaded_keys: list[str],
-) -> None:
-    """Persist uploaded part key progress so retries can resume."""
-    submission.archive_part_keys_json = json.dumps(uploaded_keys)
-    submission.last_updated_timestamp = datetime.now()
-    session.add(submission)
-    session.commit()
-
-
-def _upload_chunked_archive_parts(
-    *,
-    session: Session,
-    submission: ArchiveSubmission,
-    client: Any,
-    bucket_name: str,
-    object_prefix: str,
-    archive_parts_dir: Path,
-    timeout_seconds: int,
-) -> tuple[bool, list[str]]:
-    """Upload chunked archive part files with resume support.
-
-    A part key is considered already uploaded only if:
-    - it appears in persisted submission state, and
-    - the key currently exists in object storage.
-    """
-    part_files = sorted(archive_parts_dir.glob("*.tar.gz.part-*"))
-    uploaded_keys = _parse_uploaded_part_keys(submission.archive_part_keys_json)
-
-    for part_file in part_files:
-        part_key = f"{object_prefix}{part_file.name}"
-        if part_key in uploaded_keys:
-            exists, _ = object_exists(client, bucket_name, part_key)
-            if exists:
-                _log_event(
-                    logging.INFO,
-                    "crate.upload.part.skipped",
-                    submission_id=submission.id,
-                    drive_name=submission.drive_name,
-                    part_key=part_key,
-                    reason="already_uploaded",
-                )
-                continue
-
-        success = upload_file(
-            client,
-            bucket_name,
-            part_key,
-            file_path=str(part_file),
-            timeout=timeout_seconds,
-        )
-        if not success:
-            _log_event(
-                logging.ERROR,
-                "crate.upload.part.failed",
-                submission_id=submission.id,
-                drive_name=submission.drive_name,
-                part_key=part_key,
-            )
-            return False, uploaded_keys
-
-        uploaded_keys.append(part_key)
-        _persist_uploaded_part_keys(session, submission, uploaded_keys)
-        _log_event(
-            logging.INFO,
-            "crate.upload.part.completed",
-            submission_id=submission.id,
-            drive_name=submission.drive_name,
-            part_key=part_key,
-        )
-
-    return True, uploaded_keys
-
-
-def build_crate_contents(  # pylint: disable=too-many-arguments, too-many-positional-arguments
-    drive: dict[str, Any],
-    submission: ArchiveSubmission,
-    members_list: list[dict[str, Any]],
-    project_data: dict[str, Any],
-    drive_location: Path,
-    output_location: Path,
-) -> None:
-    """Generate RO-Crate with data from ProjectDB.
-
-    Args:
-        submission: ArchiveSubmission record for this crate generation
-        members_list: List of project members from ProjectDB
-        project_data: Project data from ProjectDB
-        drive_location: Source drive location path
-        output_location: Output archive location path
-    """
-    # Build RO-Crate
-    ro_crate_loader = ROLoader()
-    ro_crate_loader.init_crate()
-    ro_crate_builder = ROBuilder(ro_crate_loader.crate)
-
-    # Add project to crate with archive metadata
-    project_entity = ro_crate_builder.add_project(
-        project_data, members_list, submission, drive
-    )
-
-    # Add drive service as main entity
-    drive_entity = ro_crate_builder.add_research_drive_service(drive)
-
-    ro_crate_builder.crate.root_dataset.append_to("mainEntity", drive_entity)
-    drive_entity.append_to("project", [project_entity])
-
-    ro_crate_location = drive_location
-    if bagit_exists(ro_crate_location):
-        ro_crate_location = ro_crate_location / "data"
-
-    _log_event(
-        logging.INFO,
-        "crate.write.metadata",
-        ro_crate_location=str(ro_crate_location),
-        drive_name=submission.drive_name,
-    )
-    ro_crate_loader.write_crate(ro_crate_location)
-
-    _log_event(
-        logging.INFO,
-        "bag_directory.start",
-        drive_name=submission.drive_name,
-    )
-    bag_directory(
-        drive_location,
-        bag_info={
-            "project_id": str(project_data.get("id", "")),
-            "drive_name": submission.drive_name,
-        },
-    )
-
-    # Create output location after bagit processing so it doesn't get included in bag
-    output_location.mkdir(parents=True, exist_ok=True)
-
-    create_manifests_directory(
-        drive_path=drive_location,
-        output_location=output_location,
-        drive_name=str(submission.drive_name),
-    )
-
-
-async def generate_ro_crate(  # pylint: disable=too-many-locals,too-many-statements,too-many-branches
-    drive: dict[str, Any],
-    submission_id: int,
-    projectdb_client: ProjectDBClient,
-) -> None:
-    """Async background task for generating RO-Crate and updating archive record.
-
-    Fetches live project data from ProjectDB, generates crate,
-    uploads the archive to ActiveScale for long-term storage, and updates
-    the ArchiveSubmission record with stage and operational metadata.
-
-    Implements persisted checkpoints so retries can skip completed steps:
-    - queued→running: After loading submission from DB
-    - running→uploading: After crate build and tar generation
-    - uploading→completed/failed: After upload attempt
-
-    Args:
-        drive: Dictionary containing research drive information
-        submission_id: ID of the ArchiveSubmission record
-        projectdb_client: Client for interacting with ProjectDB
-    """
-    # Yield to the event loop so uvicorn can flush the HTTP response to the client
-    # before this blocking-heavy task runs.  Without this, the SelectorEventLoop on
-    # Linux holds the response in its write buffer until we return.
-    await asyncio.sleep(0)
-
-    drive_name = drive.get("name", None)
-    started_at = datetime.now()
-    if drive_name is None:
-        _log_event(
-            logging.ERROR,
-            "crate.build.invalid_drive",
-            drive=drive,
-            elapsed_ms=_elapsed_ms(started_at),
-        )
-        return
-
-    # Create a new session for this background task
-    with Session(engine) as session:
-        submission: ArchiveSubmission | None = None
-        output_location: Path | None = None
-        file_key: str | None = None
-        project_id: int | None = None
-        processing_error: str | None = None
-        upload_success = False
-        try:
-            submission = session.get(ArchiveSubmission, submission_id)
-            if submission is None:
-                _log_event(
-                    logging.ERROR,
-                    "crate.build.submission_not_found",
-                    submission_id=submission_id,
-                    drive_name=drive_name,
-                    elapsed_ms=_elapsed_ms(started_at),
-                )
-                return
-
-            # Transition: queued → packaging
-            previous_stage = submission.stage
-            submission.stage = JobStage.PACKAGING
-            submission.last_updated_timestamp = datetime.now()
-            session.add(submission)
-            session.commit()
-
-            _log_event(
-                logging.INFO,
-                "submission.stage_transition",
-                submission_id=submission_id,
-                drive_name=drive_name,
-                from_stage=previous_stage.value,
-                to_stage=JobStage.PACKAGING.value,
-                stage=submission.stage.value,
-                retry_count=submission.retry_count,
-                elapsed_ms=_elapsed_ms(started_at),
-            )
-
-            # Fetch project and member data from ProjectDB
-            project_id = submission.project_id
-            _log_event(
-                logging.INFO,
-                "crate.build.projectdb_fetch_start",
-                submission_id=submission_id,
-                drive_name=drive_name,
-                project_id=project_id,
-                stage=submission.stage.value,
-                retry_count=submission.retry_count,
-                elapsed_ms=_elapsed_ms(started_at),
-            )
-            project_data = projectdb_client.get_project(
-                pid=project_id,
-                expand=["codes", "status", "services", "properties"],
-            )
-            members_list = projectdb_client.get_project_members(
-                project_id,
-                expand=[
-                    "person",
-                    "role",
-                    "person.identities",
-                    "person.status",
-                ],
-            )
-            members_list = filter_member_identities(members_list)
-
-            # Source data and output locations
-            drive_path = _resolve_drive_path_for_archive(drive_name)
-            output_location = _resolve_archive_output_location(drive_name)
-
-            _log_event(
-                logging.INFO,
-                "crate.build.start",
-                submission_id=submission_id,
-                drive_name=drive_name,
-                project_id=project_id,
-                stage=submission.stage.value,
-                retry_count=submission.retry_count,
-                elapsed_ms=_elapsed_ms(started_at),
-            )
-
-            # Build crate contents (idempotent, safe to retry)
-            build_crate_contents(
-                drive=drive,
-                submission=submission,
-                members_list=members_list,
-                project_data=project_data,
-                drive_location=drive_path,
-                output_location=output_location,
-            )
-
-            # Build chunked tar archive package for upload.
-            settings = get_settings()
-            archive_parts_dir = output_location / "archive_parts"
-            _log_event(
-                logging.INFO,
-                "crate.package.chunked_tar.start",
-                submission_id=submission_id,
-                drive_name=drive_name,
-                archive_parts_dir=str(archive_parts_dir),
-                chunk_size_bytes=settings.archive_chunk_size_bytes,
-                stage=submission.stage.value,
-                retry_count=submission.retry_count,
-                elapsed_ms=_elapsed_ms(started_at),
-            )
-            chunk_result = build_chunked_tar_archive(
-                source_dir=drive_path,
-                output_dir=archive_parts_dir,
-                base_name=str(drive_name),
-                part_size_bytes=settings.archive_chunk_size_bytes,
-                manifest_file_name=settings.archive_chunk_manifest_file_name,
-            )
-
-            object_prefix = f"{drive_name}/"
-            submission.archive_part_count = len(chunk_result.parts)
-            submission.archive_total_bytes = chunk_result.total_bytes
-            submission.archive_object_prefix = object_prefix
-            submission.archive_manifest_key = (
-                f"{object_prefix}{chunk_result.manifest_path.name}"
-            )
-            if submission.archive_part_keys_json is None:
-                submission.archive_part_keys_json = "[]"
-            submission.last_updated_timestamp = datetime.now()
-            session.add(submission)
-            session.commit()
-
-            _log_event(
-                logging.INFO,
-                "crate.package.chunked_tar.completed",
-                submission_id=submission_id,
-                drive_name=drive_name,
-                part_count=len(chunk_result.parts),
-                total_bytes=chunk_result.total_bytes,
-                manifest_path=str(chunk_result.manifest_path),
-                stage=submission.stage.value,
-                retry_count=submission.retry_count,
-                elapsed_ms=_elapsed_ms(started_at),
-            )
-
-            # Transition: packaging → uploading
-            previous_stage = submission.stage
-            submission.stage = JobStage.UPLOADING
-            submission.last_updated_timestamp = datetime.now()
-            session.add(submission)
-            session.commit()
-
-            _log_event(
-                logging.INFO,
-                "submission.stage_transition",
-                submission_id=submission_id,
-                drive_name=drive_name,
-                from_stage=previous_stage.value,
-                to_stage=JobStage.UPLOADING.value,
-                stage=submission.stage.value,
-                retry_count=submission.retry_count,
-                elapsed_ms=_elapsed_ms(started_at),
-            )
-
-            # Upload the archive to ActiveScale
-            _log_event(
-                logging.INFO,
-                "crate.upload.start",
-                submission_id=submission_id,
-                drive_name=drive_name,
-                stage=submission.stage.value,
-                retry_count=submission.retry_count,
-                elapsed_ms=_elapsed_ms(started_at),
-            )
-
-            with get_activescale_client_context() as client:
-                bucket_name = settings.activescale_bucket_name
-
-                upload_success, uploaded_part_keys = _upload_chunked_archive_parts(
-                    session=session,
-                    submission=submission,
-                    client=client,
-                    bucket_name=bucket_name,
-                    object_prefix=object_prefix,
-                    archive_parts_dir=archive_parts_dir,
-                    timeout_seconds=settings.activescale_upload_timeout,
-                )
-
-                if upload_success:
-                    # Transition: uploading -> writing_manifest
-                    previous_stage = submission.stage
-                    submission.stage = JobStage.WRITING_MANIFEST
-                    submission.last_updated_timestamp = datetime.now()
-                    session.add(submission)
-                    session.commit()
-
-                    _log_event(
-                        logging.INFO,
-                        "submission.stage_transition",
-                        submission_id=submission_id,
-                        drive_name=drive_name,
-                        from_stage=previous_stage.value,
-                        to_stage=JobStage.WRITING_MANIFEST.value,
-                        stage=submission.stage.value,
-                        retry_count=submission.retry_count,
-                        elapsed_ms=_elapsed_ms(started_at),
-                    )
-
-                    file_key = f"{object_prefix}{chunk_result.manifest_path.name}"
-                    upload_success = upload_file(
-                        client,
-                        bucket_name,
-                        file_key,
-                        file_path=str(chunk_result.manifest_path),
-                        timeout=settings.activescale_upload_timeout,
-                        metadata={
-                            "cer_project_id": str(project_data.get("id", "")),
-                            "project_owner": get_project_owner_email(members_list),
-                            "division": project_data.get("division") or "Unknown",
-                            "data_classification": submission.data_classification
-                            or "Unknown",
-                            "retention_period_years": str(
-                                submission.retention_period_years
-                            )
-                            or "Unknown",
-                            "review_date": (
-                                calculate_retention_end_date(
-                                    datetime.now(),
-                                    submission.retention_period_years,
-                                )
-                                if submission.retention_period_years is not None
-                                else "Unknown"
-                            ),
-                            "archive_part_count": str(len(uploaded_part_keys)),
-                        },
-                    )
-                    submission.archive_manifest_key = file_key
-                    submission.archive_part_keys_json = json.dumps(uploaded_part_keys)
-                    session.add(submission)
-                    session.commit()
-
-            # Transition: uploading → cleanup
-            previous_stage = submission.stage
-            submission.stage = JobStage.CLEANUP
-            submission.last_updated_timestamp = datetime.now()
-            session.add(submission)
-            session.commit()
-
-            _log_event(
-                logging.INFO,
-                "submission.stage_transition",
-                submission_id=submission_id,
-                drive_name=drive_name,
-                from_stage=previous_stage.value,
-                to_stage=JobStage.CLEANUP.value,
-                stage=submission.stage.value,
-                retry_count=submission.retry_count,
-                elapsed_ms=_elapsed_ms(started_at),
-            )
-
-            cleanup_succeeded, cleanup_error = _cleanup_job_artifacts(
-                str(drive_name), output_location
-            )
-            submission.cleanup_succeeded = cleanup_succeeded
-            submission.cleanup_error = cleanup_error
-
-            # Update submission record with upload result
-            now = datetime.now()
-            if upload_success:
-                submission.stage = JobStage.COMPLETED
-                submission.failure_reason = None
-                submission.failed_timestamp = None
-                submission.archive_file_key = file_key
-                submission.completed_timestamp = now
-                submission.last_updated_timestamp = now
-                session.add(submission)
-                session.commit()
-
-                _log_event(
-                    logging.INFO,
-                    "crate.upload.completed",
-                    submission_id=submission_id,
-                    drive_name=drive_name,
-                    file_key=file_key,
-                    stage=submission.stage.value,
-                    retry_count=submission.retry_count,
-                    cleanup_succeeded=submission.cleanup_succeeded,
-                    elapsed_ms=_elapsed_ms(started_at),
-                )
-            else:
-                submission.stage = JobStage.FAILED
-                submission.failure_reason = "Archive upload failed"
-                submission.failed_timestamp = now
-                submission.archive_file_key = file_key
-                submission.last_updated_timestamp = now
-                session.add(submission)
-                session.commit()
-                _log_event(
-                    logging.ERROR,
-                    "crate.upload.failed",
-                    submission_id=submission_id,
-                    drive_name=drive_name,
-                    file_key=file_key,
-                    stage=submission.stage.value,
-                    retry_count=submission.retry_count,
-                    cleanup_succeeded=submission.cleanup_succeeded,
-                    elapsed_ms=_elapsed_ms(started_at),
-                )
-
-            _log_event(
-                logging.INFO,
-                "crate.build.completed",
-                submission_id=submission_id,
-                drive_name=drive_name,
-                project_id=project_id,
-                stage=submission.stage.value,
-                retry_count=submission.retry_count,
-                cleanup_succeeded=submission.cleanup_succeeded,
-                elapsed_ms=_elapsed_ms(started_at),
-            )
-        except Exception as e:  # pylint: disable=broad-exception-caught
-            processing_error = str(e)
-            if submission is not None:
-                now = datetime.now()
-
-                # Transition to cleanup before final failed state.
-                previous_stage = submission.stage
-                submission.stage = JobStage.CLEANUP
-                submission.last_updated_timestamp = now
-                session.add(submission)
-                session.commit()
-
-                _log_event(
-                    logging.WARNING,
-                    "submission.stage_transition",
-                    submission_id=submission_id,
-                    drive_name=drive_name,
-                    from_stage=previous_stage.value,
-                    to_stage=JobStage.CLEANUP.value,
-                    stage=submission.stage.value,
-                    retry_count=submission.retry_count,
-                    elapsed_ms=_elapsed_ms(started_at),
-                )
-
-                cleanup_succeeded, cleanup_error = _cleanup_job_artifacts(
-                    str(drive_name), output_location
-                )
-                submission.cleanup_succeeded = cleanup_succeeded
-                submission.cleanup_error = cleanup_error
-
-                submission.stage = JobStage.FAILED
-                submission.failure_reason = processing_error
-                submission.failed_timestamp = now
-                submission.last_updated_timestamp = now
-                session.add(submission)
-                session.commit()
-            _log_event(
-                logging.ERROR,
-                "crate.build.failed",
-                submission_id=submission_id,
-                drive_name=drive_name,
-                error=processing_error,
-                stage=(submission.stage.value if submission is not None else None),
-                retry_count=(
-                    submission.retry_count if submission is not None else None
-                ),
-                cleanup_succeeded=(
-                    submission.cleanup_succeeded if submission is not None else None
-                ),
-                elapsed_ms=_elapsed_ms(started_at),
-            )
-            _log_event(
-                logging.ERROR,
-                "crate.build.exception",
-                submission_id=submission_id,
-                drive_name=drive_name,
-                exc_info=True,
-                elapsed_ms=_elapsed_ms(started_at),
-            )
-
-
-@app.get(
-    ENDPOINT_PREFIX + "/submission",
-    response_model=SubmissionResponse,
-    responses={
-        401: {"model": ErrorResponse, "description": "Invalid or missing API key"},
-        404: {
-            "model": ErrorResponse,
-            "description": "No archive submission found for drive",
-        },
-    },
-)
-async def get_submission(
-    drive_name: ResearchDriveName,
-    session: SessionDep,
-    api_key: ApiKey = Security(validate_api_key),
-) -> SubmissionResponse:
-    """Retrieve archive submission record for a research drive."""
-    validate_permissions("GET", api_key)
-
-    stmt = select(ArchiveSubmission).where(ArchiveSubmission.drive_name == drive_name)
-    submission = session.exec(stmt).first()
-
-    if submission is None:
-        raise HTTPException(
-            status_code=404,
-            detail=f"No archive submission found for drive {drive_name}",
-        )
-
-    return SubmissionResponse(
-        drive_id=submission.drive_id,
-        project_id=submission.project_id,
-        drive_name=submission.drive_name,
-        retention_period_years=submission.retention_period_years,
-        retention_period_justification=submission.retention_period_justification,
-        data_classification=submission.data_classification,
-        stage=submission.stage,
-        failure_reason=submission.failure_reason,
-        failed_timestamp=submission.failed_timestamp,
-        started_timestamp=submission.started_timestamp,
-        last_updated_timestamp=submission.last_updated_timestamp,
-        completed_timestamp=submission.completed_timestamp,
-        retry_count=submission.retry_count,
-        cleanup_succeeded=submission.cleanup_succeeded,
-        cleanup_error=submission.cleanup_error,
-        archive_file_key=submission.archive_file_key,
-        archive_object_prefix=submission.archive_object_prefix,
-        archive_manifest_key=submission.archive_manifest_key,
-        archive_part_keys_json=submission.archive_part_keys_json,
-        archive_part_count=submission.archive_part_count,
-        archive_total_bytes=submission.archive_total_bytes,
-    )
-
-
-def _build_codes(project_data: dict[str, Any]) -> list[CodeResponse]:
-    """Extract project codes from project data."""
-    codes_items = project_data.get("codes", {})
-    if isinstance(codes_items, dict):
-        codes_items = codes_items.get("items", [])
-    return [CodeResponse(id=c.get("id"), code=c["code"]) for c in codes_items]
-
-
-def _build_members(members_raw: list[dict[str, Any]]) -> list[MemberResponse]:
-    """Convert raw member dicts into MemberResponse objects."""
-    members = []
-    for m in members_raw:
-        person = m.get("person", {})
-        username = None
-        for ident in person.get("identities", {}).get("items", []):
-            uname = ident.get("username", "")
-            if uname and "@" not in uname:
-                username = uname
-                break
-
-        members.append(
-            MemberResponse(
-                role=RoleResponse(
-                    id=m.get("role", {}).get("id"),
-                    name=m["role"]["name"],
-                ),
-                person=PersonResponse(
-                    id=person.get("id"),
-                    email=person.get("email"),
-                    full_name=person.get("full_name", ""),
-                    username=username,
-                ),
-            )
-        )
-    return members
-
-
-def filter_member_identities(members: list[dict[str, Any]]) -> list[dict[str, Any]]:
-    """Filter out identities where the username is an email address."""
-    try:
-        members = [
-            {
-                **member,
-                "person": {
-                    **member.get("person", {}),
-                    "identities": {
-                        "items": [
-                            item
-                            for item in member.get("person", {})
-                            .get("identities", {})
-                            .get("items", [])
-                            if not item.get("username", "").endswith("@auckland.ac.nz")
-                        ]
-                    },
-                },
-            }
-            for member in members
-        ]
-    except (TypeError, AttributeError) as e:
-        # Log error but don't fail the whole process - just return unfiltered members
-        _log_event(logging.WARNING, "members.filter_failed", error=str(e))
-    return members
-
-
-def get_project_owner_email(members: Any) -> str:
-    """Get the project owner's email from the members list."""
-    for member in members:
-        if member["role"]["name"] == "Project Owner":
-            return member["person"]["email"] or "Unknown"
-    return "Unknown"
-
-
-def _transition_retrieval_stage(
-    session: Session,
-    retrieval: ArchiveRetrieval,
-    to_stage: RetrievalJobStage,
-    started_at: datetime,
-) -> None:
-    """Transition a retrieval job to *to_stage*, commit, and log the change."""
-    previous_stage = retrieval.stage
-    now = datetime.now()
-    if to_stage == RetrievalJobStage.RESTORING and retrieval.started_timestamp is None:
-        retrieval.started_timestamp = now
-    retrieval.stage = to_stage
-    retrieval.last_updated_timestamp = now
-    session.add(retrieval)
-    session.commit()
-    _log_event(
-        logging.INFO,
-        "retrieval.stage_transition",
-        retrieval_id=retrieval.id,
-        drive_name=retrieval.drive_name,
-        from_stage=previous_stage.value,
-        to_stage=to_stage.value,
-        elapsed_ms=_elapsed_ms(started_at),
-    )
-
-
-def _parse_retrieved_part_keys(part_keys_json: str | None) -> list[str]:
-    """Decode the JSON-encoded list of already-downloaded part object keys."""
-    if not part_keys_json:
-        return []
-    try:
-        parsed = json.loads(part_keys_json)
-    except json.JSONDecodeError:
-        return []
-    if not isinstance(parsed, list):
-        return []
-    return [str(item) for item in parsed if isinstance(item, str)]
-
-
-def _persist_retrieved_part_keys(
-    session: Session,
-    retrieval: ArchiveRetrieval,
-    downloaded_keys: list[str],
-) -> None:
-    """Persist the current list of downloaded part keys to the retrieval record."""
-    retrieval.retrieved_part_keys_json = json.dumps(downloaded_keys)
-    retrieval.last_updated_timestamp = datetime.now()
-    session.add(retrieval)
-    session.commit()
-
-
-async def run_archive_retrieval(retrieval_id: int) -> None:  # pylint: disable=too-many-statements
-    """Background task: restore, download, and extract a completed archive.
-
-    Workflow:
-      1. RESTORING  - Download the manifest from S3; initiate a restore request
-                      for every archive part; poll until all parts are available.
-      2. DOWNLOADING - Download each archive part to a local temp directory,
-                      persisting progress so a future retry can resume mid-download.
-      3. EXTRACTING  - Reassemble the chunked parts into a single .tar.gz; extract
-                      into destination_path; validate the resulting BagIt bag;
-                      clean up temp files.
-      4. COMPLETED / FAILED - Final state written to the ArchiveRetrieval record.
-    """
-    # Yield to the event loop so uvicorn can flush the HTTP response to the client
-    # before this blocking-heavy task begins.
-    await asyncio.sleep(0)
-
-    started_at = datetime.now()
-    settings = get_settings()
-
-    with Session(engine) as session:
-        retrieval: ArchiveRetrieval | None = None
-        download_dir: Path | None = None
-
-        try:
-            retrieval = session.get(ArchiveRetrieval, retrieval_id)
-            if retrieval is None:
-                _log_event(
-                    logging.ERROR,
-                    "retrieval.run.record_not_found",
-                    retrieval_id=retrieval_id,
-                )
-                return
-
-            submission = session.get(ArchiveSubmission, retrieval.submission_id)
-            if submission is None:
-                raise RuntimeError(
-                    f"ArchiveSubmission {retrieval.submission_id} not found in database"
-                )
-
-            drive_name = submission.drive_name
-            manifest_key = submission.archive_manifest_key
-            if not manifest_key:
-                raise RuntimeError(
-                    "Submission has no archive_manifest_key; archive may not have"
-                    " uploaded completely"
-                )
-
-            object_prefix = submission.archive_object_prefix or f"{drive_name}/"
-            bucket_name = settings.activescale_bucket_name
-
-            # Isolated temp directory for this retrieval job (keyed by ID so
-            # concurrent retrievals never share scratch space).
-            temp_base = Path(settings.archive_temp_base_path).expanduser()
-            download_dir = temp_base / "retrieval" / f"{retrieval_id}_{drive_name}"
-            download_dir.mkdir(parents=True, exist_ok=True)
-
-            _log_event(
-                logging.INFO,
-                "retrieval.run.started",
-                retrieval_id=retrieval_id,
-                drive_name=drive_name,
-                destination_path=retrieval.destination_path,
-                bucket_name=bucket_name,
-            )
-
-            # ─── Phase 1: RESTORING ───────────────────────────────────────────
-            _transition_retrieval_stage(
-                session, retrieval, RetrievalJobStage.RESTORING, started_at
-            )
-
-            manifest_local = download_dir / settings.archive_chunk_manifest_file_name
-            poll_interval = settings.activescale_restore_poll_interval_seconds
-            max_wait = settings.activescale_restore_poll_max_seconds
-
-            # Step 1a: Restore and download the manifest.
-            # The manifest itself may be on tape and require a restore before
-            # it can be read.  We only need to do this if it has not already
-            # been downloaded to the local scratch directory (e.g. on a resume).
-            if not manifest_local.exists():
-                with get_activescale_client_context() as client:
-                    manifest_needs_restore = initiate_object_restore(
-                        client, bucket_name, manifest_key,
-                        days=settings.activescale_restore_days,
-                    )
-
-                if manifest_needs_restore:
-                    deadline = datetime.now() + timedelta(seconds=max_wait)
-                    while True:
-                        with get_activescale_client_context() as client:
-                            ready = is_object_ready_for_download(
-                                client, bucket_name, manifest_key
-                            )
-                        if ready:
-                            break
-                        if datetime.now() >= deadline:
-                            raise TimeoutError(
-                                f"Archive manifest not restored within {max_wait}s."
-                                " Re-submit a retrieval request to try again."
-                            )
-                        _log_event(
-                            logging.INFO,
-                            "retrieval.restore.manifest.waiting",
-                            retrieval_id=retrieval_id,
-                            drive_name=drive_name,
-                            manifest_key=manifest_key,
-                            poll_interval_seconds=poll_interval,
-                            elapsed_ms=_elapsed_ms(started_at),
-                        )
-                        await asyncio.sleep(poll_interval)
-
-                with get_activescale_client_context() as client:
-                    if not download_file_to_disk(
-                        client, bucket_name, manifest_key, manifest_local
-                    ):
-                        raise RuntimeError(
-                            f"Failed to download archive manifest: {manifest_key}"
-                        )
-
-            manifest_data = load_archive_manifest(manifest_local)
-            part_keys = ordered_part_object_keys(object_prefix, manifest_data)
-
-            # Step 1b: Restore archive parts.
-            _log_event(
-                logging.INFO,
-                "retrieval.restore.initiating",
-                retrieval_id=retrieval_id,
-                drive_name=drive_name,
-                part_count=len(part_keys),
-                bucket_name=bucket_name,
-            )
-
-            # Send restore requests; track only parts that are on tape
-            # (parts already in active storage skip the polling loop).
-            needs_restore: list[str] = []
-            with get_activescale_client_context() as client:
-                for part_key in part_keys:
-                    if initiate_object_restore(
-                        client, bucket_name, part_key,
-                        days=settings.activescale_restore_days,
-                    ):
-                        needs_restore.append(part_key)
-
-            # Poll until all tape-tier parts have been thawed.
-            if needs_restore:
-                deadline = datetime.now() + timedelta(seconds=max_wait)
-
-                while True:
-                    with get_activescale_client_context() as client:
-                        all_ready = all(
-                            is_object_ready_for_download(client, bucket_name, k)
-                            for k in needs_restore
-                        )
-                    if all_ready:
-                        _log_event(
-                            logging.INFO,
-                            "retrieval.restore.complete",
-                            retrieval_id=retrieval_id,
-                            drive_name=drive_name,
-                            elapsed_ms=_elapsed_ms(started_at),
-                        )
-                        break
-                    if datetime.now() >= deadline:
-                        raise TimeoutError(
-                            f"Archive objects not restored within {max_wait}s."
-                            " Re-submit a retrieval request to try again."
-                        )
-                    _log_event(
-                        logging.INFO,
-                        "retrieval.restore.polling",
-                        retrieval_id=retrieval_id,
-                        drive_name=drive_name,
-                        pending_count=len(needs_restore),
-                        poll_interval_seconds=poll_interval,
-                        elapsed_ms=_elapsed_ms(started_at),
-                    )
-                    await asyncio.sleep(poll_interval)
-
-            # ─── Phase 2: DOWNLOADING ─────────────────────────────────────────
-            _transition_retrieval_stage(
-                session, retrieval, RetrievalJobStage.DOWNLOADING, started_at
-            )
-
-            already_downloaded = _parse_retrieved_part_keys(
-                retrieval.retrieved_part_keys_json
-            )
-
-            with get_activescale_client_context() as client:
-                for part_key in part_keys:
-                    if part_key in already_downloaded:
-                        _log_event(
-                            logging.DEBUG,
-                            "retrieval.download.part.skip",
-                            retrieval_id=retrieval_id,
-                            part_key=part_key,
-                        )
-                        continue
-
-                    # Derive local filename from the trailing segment of the object key.
-                    part_filename = part_key.removeprefix(object_prefix)
-                    dest = download_dir / part_filename
-
-                    _log_event(
-                        logging.INFO,
-                        "retrieval.download.part.start",
-                        retrieval_id=retrieval_id,
-                        part_key=part_key,
-                        dest=str(dest),
-                        elapsed_ms=_elapsed_ms(started_at),
-                    )
-
-                    if not download_file_to_disk(client, bucket_name, part_key, dest):
-                        raise RuntimeError(
-                            f"Failed to download archive part: {part_key}"
-                        )
-
-                    already_downloaded.append(part_key)
-                    _persist_retrieved_part_keys(session, retrieval, already_downloaded)
-
-            # ─── Phase 3: EXTRACTING ──────────────────────────────────────────
-            _transition_retrieval_stage(
-                session, retrieval, RetrievalJobStage.EXTRACTING, started_at
-            )
-
-            reassembled_tar = download_dir / f"{drive_name}.tar.gz"
-            reassemble_archive_from_manifest(
-                parts_dir=download_dir,
-                manifest_path=manifest_local,
-                output_tar_path=reassembled_tar,
-                verify_parts=True,
-            )
-
-            _log_event(
-                logging.INFO,
-                "retrieval.extract.start",
-                retrieval_id=retrieval_id,
-                drive_name=drive_name,
-                tar_path=str(reassembled_tar),
-                elapsed_ms=_elapsed_ms(started_at),
-            )
-
-            dest_path = Path(retrieval.destination_path)
-            with tarfile.open(reassembled_tar, "r:gz") as tar:
-                tar.extractall(path=dest_path, filter="data")
-
-            # Validate BagIt integrity of the extracted archive.
-            # ``source_root`` from the manifest tells us the top-level directory
-            # name that was stored in the tar (mirrors tarfile.add arcname).
-            source_root = str(manifest_data.get("source_root", drive_name))
-            extracted_bag_path = dest_path / source_root
-            if bagit_exists(extracted_bag_path):
-                validate_bag(extracted_bag_path)
-                _log_event(
-                    logging.INFO,
-                    "retrieval.bagit.validated",
-                    retrieval_id=retrieval_id,
-                    bag_path=str(extracted_bag_path),
-                    elapsed_ms=_elapsed_ms(started_at),
-                )
-            else:
-                _log_event(
-                    logging.WARNING,
-                    "retrieval.bagit.not_found",
-                    retrieval_id=retrieval_id,
-                    expected_path=str(extracted_bag_path),
-                )
-
-            # Clean up scratch space now that extraction has been verified.
-            shutil.rmtree(download_dir, ignore_errors=True)
-            download_dir = None
-
-            # Mark COMPLETED
-            now = datetime.now()
-            retrieval.stage = RetrievalJobStage.COMPLETED
-            retrieval.completed_timestamp = now
-            retrieval.last_updated_timestamp = now
-            session.add(retrieval)
-            session.commit()
-
-            _log_event(
-                logging.INFO,
-                "retrieval.completed",
-                retrieval_id=retrieval_id,
-                drive_name=drive_name,
-                destination_path=retrieval.destination_path,
-                elapsed_ms=_elapsed_ms(started_at),
-            )
-
-        except Exception as e:  # pylint: disable=broad-exception-caught
-            if download_dir is not None:
-                shutil.rmtree(download_dir, ignore_errors=True)
-
-            if retrieval is not None:
-                now = datetime.now()
-                retrieval.stage = RetrievalJobStage.FAILED
-                retrieval.failure_reason = str(e)
-                retrieval.failed_timestamp = now
-                retrieval.last_updated_timestamp = now
-                session.add(retrieval)
-                session.commit()
-
-            _log_event(
-                logging.ERROR,
-                "retrieval.failed",
-                retrieval_id=retrieval_id,
-                error_type=type(e).__name__,
-                error=str(e),
-                exc_info=True,
-            )
-
-
-@app.post(
-    ENDPOINT_PREFIX + "/submission/{drive_name}/retrieve",
-    status_code=status.HTTP_201_CREATED,
-    response_model=CreateRetrievalResponse,
-    responses={
-        400: {"model": ErrorResponse, "description": "Invalid retrieval request"},
-        401: {"model": ErrorResponse, "description": "Invalid or missing API key"},
-        404: {
-            "model": ErrorResponse,
-            "description": "No completed archive submission found for drive",
-        },
-        409: {
-            "model": ErrorResponse,
-            "description": "A retrieval job for this drive is already active",
-        },
-        422: {"description": "Validation error"},
-        500: {"model": ErrorResponse, "description": "Internal server error"},
-    },
-)
-async def create_retrieval(
-    drive_name: ResearchDriveName,
-    request: CreateRetrievalRequest,
-    session: SessionDep,
-    background_tasks: BackgroundTasks,
-    api_key: ApiKey = Security(validate_api_key),
-) -> CreateRetrievalResponse:
-    """Schedule an archive retrieval job for a research drive.
-
-    Validates that a completed archive exists for the drive and that the
-    destination path is accessible, then schedules a background task to
-    restore, download, and extract the archive into the destination.
-    """
-    validate_permissions("POST", api_key)
-
-    # 1. Verify a completed archive submission exists for this drive.
-    submission = session.exec(
-        select(ArchiveSubmission).where(ArchiveSubmission.drive_name == drive_name)
-    ).first()
-
-    if submission is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"No archive submission found for drive {drive_name}.",
-        )
-
-    if submission.stage != JobStage.COMPLETED:
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail=(
-                f"Archive for drive {drive_name} is not in a completed state"
-                f" (current stage: '{submission.stage.value}')."
-                " Retrieval requires a successfully completed archive."
-            ),
-        )
-
-    if not submission.archive_manifest_key:
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail=(
-                f"Archive submission for drive {drive_name} has no manifest key."
-                " The archive may be incomplete."
-            ),
-        )
-
-    # 2. Check there is no active retrieval job already running for this drive.
-    active_stage_values = [s.value for s in ACTIVE_RETRIEVAL_STAGES]
-    stage_column = cast(Any, ArchiveRetrieval.stage)
-    active_retrieval = session.exec(
-        select(ArchiveRetrieval).where(
-            ArchiveRetrieval.drive_name == drive_name,
-            stage_column.in_(active_stage_values),
-        )
-    ).first()
-
-    if active_retrieval is not None:
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail=(
-                f"A retrieval job for drive {drive_name} is already active"
-                f" (stage: '{active_retrieval.stage.value}')."
-            ),
-        )
-
-    # 3. Validate the destination path exists and is writable.
-    try:
-        _validate_destination_path(request.destination_path)
-    except (FileNotFoundError, PermissionError) as e:
-        _log_event(
-            logging.WARNING,
-            "retrieval.destination_path_validation_failed",
-            drive_name=drive_name,
-            destination_path=request.destination_path,
-            error=str(e),
-            error_type=type(e).__name__,
-        )
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=f"Destination path validation failed: {e}",
-        ) from e
-
-    # 4. Create the retrieval record.
-    if submission.id is None:
-        raise HTTPException(
-            status_code=500,
-            detail="Archive submission record is missing ID.",
-        )
-
-    now = datetime.now()
-    retrieval = ArchiveRetrieval(
-        drive_name=drive_name,
-        submission_id=submission.id,
-        destination_path=request.destination_path,
-        stage=RetrievalJobStage.QUEUED,
-        started_timestamp=now,
-        last_updated_timestamp=now,
-    )
-    session.add(retrieval)
-    session.commit()
-    session.refresh(retrieval)
-
-    if retrieval.id is None:
-        raise HTTPException(
-            status_code=500,
-            detail="Failed to create archive retrieval record.",
-        )
-
-    _log_event(
-        logging.INFO,
-        "retrieval.background_task_scheduled",
-        retrieval_id=retrieval.id,
-        drive_name=drive_name,
-        destination_path=request.destination_path,
-        submission_id=submission.id,
-    )
-
-    # 5. Schedule the background task.
-    background_tasks.add_task(run_archive_retrieval, retrieval.id)
-
-    return CreateRetrievalResponse(
-        message=(
-            f"Archive retrieval job created for {drive_name}."
-            f" Restoring archive to {request.destination_path}."
-        )
-    )
+app.include_router(drives.router, prefix=ENDPOINT_PREFIX)
+app.include_router(submissions.router, prefix=ENDPOINT_PREFIX)
+app.include_router(retrievals.router, prefix=ENDPOINT_PREFIX)
