+++ src/packaging/manifests.py
